requirements.txt
# ML & Data
pandas
numpy
scikit-learn
mlflow
psycopg2-binary
kagglehub
matplotlib
imblearn
pandantic

# API
fastapi
uvicorn
python-dotenv

# Airflow Providers (нужны для тасков)
apache-airflow-providers-docker
apache-airflow-providers-postgres
